dev-requirements.txt
flake8
pytest
pytest-django
tox
